// src/types/generated/PromptVersion.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export interface PromptVersion { id: number | null, templateId: number, versionNumber: number, isActive: boolean, content: string, metadata: string | null, createdBy: string, createdAt: string, }