// src/types/generated/PromptVersionDiff.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ComponentDiff } from "./ComponentDiff";
import type { MetadataDiff } from "./MetadataDiff";
import type { ParameterDiff } from "./ParameterDiff";
import type { PromptVersion } from "./PromptVersion";

export interface PromptVersionDiff { fromVersion: PromptVersion, toVersion: PromptVersion, componentChanges: Array<ComponentDiff>, parameterChanges: Array<ParameterDiff>, metadataChanges: MetadataDiff | null, }